+++ test/testcases/utils/engine_utils.py
@@ -20,7 +20,7 @@ _DOC_ENGINE_CACHE = None
 
 
 def get_doc_engine(rag=None) -> str:
-    """Return lower-cased doc_engine from env, or from /system/status if env is unset."""
+    """Return lower-cased doc_engine from env, or from /api/v1/system/status if env is unset."""
     global _DOC_ENGINE_CACHE
     env = (os.getenv("DOC_ENGINE") or "").strip().lower()
     if env:
@@ -34,9 +34,9 @@ def get_doc_engine(rag=None) -> str:
         api_url = getattr(rag, "api_url", "")
         if "/api/" in api_url:
             base_url, version = api_url.rsplit("/api/", 1)
-            status_url = f"{base_url}/{version}/system/status"
+            status_url = f"{base_url}/api/{version}/system/status"
         else:
-            status_url = f"{api_url}/system/status"
+            status_url = f"{api_url}/api/v1/system/status"
         headers = getattr(rag, "authorization_header", {})
         res = requests.get(status_url, headers=headers).json()
         engine = str(res.get("data", {}).get("doc_engine", {}).get("type", "")).lower()
