+++ web/src/utils/api.ts
@@ -172,7 +172,7 @@ export default {
   getSystemTokenList: `${restAPIv1}/system/tokens`,
   createSystemToken: `${restAPIv1}/system/tokens`,
   removeSystemToken: `${restAPIv1}/system/tokens`,
-  getSystemConfig: `${webAPI}/system/config`,
+  getSystemConfig: `${restAPIv1}/system/config`,
   setLangfuseConfig: `${webAPI}/langfuse/api_key`,
 
   // flow
